Added support for setting window position to given screen coordinates.
Newly introduced AppWatchers regularly check the application and perform some action, for example setting the window position and size. Former AppInitDetector now also based on AppWatchers. The application launched is now considered initialized from the very beginning. But the AppWatcher can immediately clear the Initialized flag and postpone setting it back until given condition is met. Then the ask the system to stop ticking and to remove the AppWatcher.

WindowPositioner shows another usage of the AppWatcher concept. It scans the windows belonging to the process. If it finds one with matching title it sets the position and size of the window and stops operating.